requirements.txt
boto3==1.34.77
pandas==2.2.2
requests==2.31.0
python-dotenv==1.0.1
